Feature: Allow dotfiles to be linked and parsed
There are currently methods to allow for a user to link their dotfiles
from ~/.dotfiles to ~

There is also support for skipping lines and marking lines as private.
The goal of private lines is to allow you to keep your environment
variables and private data in a public git repository.

TODO:
 * Parse the files on a pull to allow them to be changed to the
actual values easily.
 * Allow dotfiles to be created via a git clone from a github user.